requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.3
async-timeout==4.0.2
attrs==23.1.0
awscli @ file:///Users/kirstenhunter/src/bedrock-python-sdk/awscli-1.29.21-py3-none-any.whl#sha256=359bb5da1692c7085e7eed3ce752670339f6beb211dd76bc51c6e75bb549410d
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
boto3 @ file:///Users/kirstenhunter/src/bedrock-python-sdk/boto3-1.28.21-py3-none-any.whl#sha256=fd7ac97960fb52106759c6a072ffcda8d106c9f75e7d8301fe1867048369b56e
botocore @ file:///Users/kirstenhunter/src/bedrock-python-sdk/botocore-1.31.21-py3-none-any.whl#sha256=96bc05a853e9bc4a385c50cf8733997ebd716ce56cc94c0b3585a26781423f66
cachetools==5.3.1
cassandra-driver==3.28.0
cassio==0.1.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.4
comm==0.1.3
contextlib2==21.6.0
dataclasses-json==0.5.13
datasets==2.14.4
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
dnspython==2.4.2
docutils==0.16
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
Faker==19.6.1
fastjsonschema==2.18.0
filelock==3.12.2
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.4.0
fs==2.4.16
fsspec==2023.6.0
gast==0.4.0
geomet==0.2.1.post1
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.28.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
gql==3.4.1
graphql-core==3.2.3
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.2
grpcio-status==1.56.2
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isoduration==20.11.0
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.12.0
langchain==0.0.249
langchainplus-sdk==0.0.20
langsmith==0.0.40
libclang==16.0.6
loguru==0.7.0
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mindsdb-sdk==1.0.5
mindsdb-sql==0.6.7
mistune==3.0.1
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==7.0.0
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
octohub==0.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pinecone-client==2.2.2
platformdirs==3.9.1
pluggy==1.2.0
pox==0.3.3
ppft==1.7.6.7
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.7.2
s3transfer==0.6.2
safetensors==0.3.1
sagemaker==2.185.0
schema==0.7.5
scikit-learn==1.3.0
scipy==1.11.1
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
Shapely==1.8.5.post1
six==1.16.0
sly==0.5
smdebug-rulesconfig==1.0.1
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
stack-data==0.6.2
sympy==1.12
tblib==1.7.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow-cpu==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==3.0.8
wrapt==1.14.1
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
